move login/register/logout code to SessionPair
